move everything to lots of subfolders
