+++ src/battle/ActivePokemon.java
@@ -23,9 +23,9 @@ import battle.effect.generic.EffectInterfaces.NameChanger;
 import battle.effect.generic.EffectInterfaces.OpponentItemBlockerEffect;
 import battle.effect.generic.EffectInterfaces.OpponentTrappingEffect;
 import battle.effect.generic.EffectInterfaces.TrappingEffect;
-import battle.effect.generic.PokemonEffect;
-import battle.effect.generic.PokemonEffectNamesies;
-import battle.effect.generic.TeamEffectNamesies;
+import battle.effect.generic.pokemon.PokemonEffect;
+import battle.effect.generic.pokemon.PokemonEffectNamesies;
+import battle.effect.generic.team.TeamEffectNamesies;
 import battle.effect.holder.AbilityHolder;
 import battle.effect.holder.ItemHolder;
 import battle.effect.status.Status;
