+++ src/battle/Battle.java
@@ -4,7 +4,7 @@ import battle.attack.Attack;
 import battle.attack.MoveType;
 import battle.effect.InvokeEffect;
 import battle.effect.attack.MultiTurnMove;
-import battle.effect.generic.BattleEffect;
+import battle.effect.generic.battle.BattleEffect;
 import battle.effect.generic.EffectInterfaces.AlwaysCritEffect;
 import battle.effect.generic.EffectInterfaces.BasicAccuracyBypassEffect;
 import battle.effect.generic.EffectInterfaces.BeforeTurnEffect;
@@ -22,8 +22,8 @@ import battle.effect.generic.EffectInterfaces.SemiInvulnerableBypasser;
 import battle.effect.generic.EffectInterfaces.StallingEffect;
 import battle.effect.generic.EffectInterfaces.StrikeFirstEffect;
 import battle.effect.generic.EffectNamesies.BattleEffectNamesies;
-import battle.effect.generic.StandardBattleEffectNamesies;
-import battle.effect.generic.WeatherEffect;
+import battle.effect.generic.battle.StandardBattleEffectNamesies;
+import battle.effect.generic.battle.weather.WeatherEffect;
 import main.Game;
 import main.Global;
 import map.overworld.TerrainType;
