+++ src/battle/attack/Attack.java
@@ -44,12 +44,12 @@ import battle.effect.generic.EffectInterfaces.StatSwitchingEffect;
 import battle.effect.generic.EffectInterfaces.SwapOpponentEffect;
 import battle.effect.generic.EffectInterfaces.TargetSwapperEffect;
 import battle.effect.generic.EffectNamesies;
-import battle.effect.generic.PokemonEffect;
-import battle.effect.generic.PokemonEffectNamesies;
-import battle.effect.generic.StandardBattleEffectNamesies;
-import battle.effect.generic.TeamEffectNamesies;
-import battle.effect.generic.TerrainNamesies;
-import battle.effect.generic.WeatherNamesies;
+import battle.effect.generic.pokemon.PokemonEffect;
+import battle.effect.generic.pokemon.PokemonEffectNamesies;
+import battle.effect.generic.battle.StandardBattleEffectNamesies;
+import battle.effect.generic.team.TeamEffectNamesies;
+import battle.effect.generic.battle.terrain.TerrainNamesies;
+import battle.effect.generic.battle.weather.WeatherNamesies;
 import battle.effect.holder.ItemHolder;
 import battle.effect.status.Status;
 import battle.effect.status.StatusCondition;
