+++ src/battle/effect/generic/EffectInterfaces.java
@@ -9,7 +9,13 @@ import battle.attack.MoveType;
 import battle.effect.CastSource;
 import battle.effect.InvokeEffect;
 import battle.effect.attack.MultiTurnMove;
+import battle.effect.generic.battle.BattleEffect;
+import battle.effect.generic.battle.weather.WeatherEffect;
+import battle.effect.generic.battle.weather.WeatherNamesies;
 import battle.effect.generic.EffectNamesies.BattleEffectNamesies;
+import battle.effect.generic.pokemon.PokemonEffect;
+import battle.effect.generic.pokemon.PokemonEffectNamesies;
+import battle.effect.generic.team.TeamEffect;
 import battle.effect.status.StatusCondition;
 import item.Item;
 import item.ItemNamesies;
