+++ src/battle/effect/generic/battle/terrain/TerrainNamesies.java
@@ -1,10 +1,10 @@
-package battle.effect.generic;
+package battle.effect.generic.battle.terrain;
 
+import battle.effect.generic.battle.terrain.TerrainEffect.ElectricTerrain;
+import battle.effect.generic.battle.terrain.TerrainEffect.GrassyTerrain;
+import battle.effect.generic.battle.terrain.TerrainEffect.MistyTerrain;
+import battle.effect.generic.battle.terrain.TerrainEffect.PsychicTerrain;
 import battle.effect.generic.EffectNamesies.BattleEffectNamesies;
-import battle.effect.generic.TerrainEffect.ElectricTerrain;
-import battle.effect.generic.TerrainEffect.GrassyTerrain;
-import battle.effect.generic.TerrainEffect.MistyTerrain;
-import battle.effect.generic.TerrainEffect.PsychicTerrain;
 
 import java.util.function.Supplier;
