+++ src/generator/NamesiesType.java
@@ -1,11 +1,11 @@
 package generator;
 
 import battle.attack.AttackNamesies;
-import battle.effect.generic.StandardBattleEffectNamesies;
-import battle.effect.generic.PokemonEffectNamesies;
-import battle.effect.generic.TeamEffectNamesies;
-import battle.effect.generic.TerrainNamesies;
-import battle.effect.generic.WeatherNamesies;
+import battle.effect.generic.battle.StandardBattleEffectNamesies;
+import battle.effect.generic.pokemon.PokemonEffectNamesies;
+import battle.effect.generic.team.TeamEffectNamesies;
+import battle.effect.generic.battle.terrain.TerrainNamesies;
+import battle.effect.generic.battle.weather.WeatherNamesies;
 import item.ItemNamesies;
 import pokemon.PokemonNamesies;
 import pokemon.ability.AbilityNamesies;
@@ -13,11 +13,11 @@ import util.Folder;
 
 public enum NamesiesType {
     ATTACK_NAMESIES(AttackNamesies.class, Folder.ATTACK, true),
-    POKEMON_EFFECT_NAMESIES(PokemonEffectNamesies.class, Folder.GENERIC_EFFECT, false),
-    TEAM_EFFECT_NAMESIES(TeamEffectNamesies.class, Folder.GENERIC_EFFECT, false),
-    BATTLE_EFFECT_NAMESIES(StandardBattleEffectNamesies.class, Folder.GENERIC_EFFECT, false),
-    WEATHER_NAMESIES(WeatherNamesies.class, Folder.GENERIC_EFFECT, false),
-    TERRAIN_NAMESIES(TerrainNamesies.class, Folder.GENERIC_EFFECT, false),
+    POKEMON_EFFECT_NAMESIES(PokemonEffectNamesies.class, Folder.POKEMON_EFFECT, false),
+    TEAM_EFFECT_NAMESIES(TeamEffectNamesies.class, Folder.TEAM_EFFECT, false),
+    BATTLE_EFFECT_NAMESIES(StandardBattleEffectNamesies.class, Folder.BATTLE_EFFECT, false),
+    WEATHER_NAMESIES(WeatherNamesies.class, Folder.WEATHER_EFFECT, false),
+    TERRAIN_NAMESIES(TerrainNamesies.class, Folder.TERRAIN_EFFECT, false),
     ABILITY_NAMESIES(AbilityNamesies.class, Folder.ABILITY, true),
     ITEM_NAMESIES(ItemNamesies.class, Folder.ITEMS, true),
     POKEMON_NAMESIES(PokemonNamesies.class, Folder.POKEMON, true);
