+++ src/item/Item.java
@@ -40,9 +40,9 @@ import battle.effect.generic.EffectInterfaces.TerrainCastEffect;
 import battle.effect.generic.EffectInterfaces.WeatherBlockerEffect;
 import battle.effect.generic.EffectInterfaces.WeatherExtendingEffect;
 import battle.effect.generic.EffectNamesies;
-import battle.effect.generic.PokemonEffectNamesies;
-import battle.effect.generic.TeamEffectNamesies;
-import battle.effect.generic.WeatherNamesies;
+import battle.effect.generic.pokemon.PokemonEffectNamesies;
+import battle.effect.generic.team.TeamEffectNamesies;
+import battle.effect.generic.battle.weather.WeatherNamesies;
 import battle.effect.status.Status;
 import battle.effect.status.StatusCondition;
 import item.bag.BagCategory;
