+++ src/map/overworld/TerrainType.java
@@ -2,7 +2,7 @@ package map.overworld;
 
 import battle.attack.AttackNamesies;
 import battle.effect.generic.EffectNamesies;
-import battle.effect.generic.PokemonEffectNamesies;
+import battle.effect.generic.pokemon.PokemonEffectNamesies;
 import battle.effect.status.StatusCondition;
 import main.Game;
 import pokemon.Stat;
