+++ src/pokemon/ability/Ability.java
@@ -76,11 +76,10 @@ import battle.effect.generic.EffectInterfaces.WeatherBlockerEffect;
 import battle.effect.generic.EffectInterfaces.WeatherEliminatingEffect;
 import battle.effect.generic.EffectInterfaces.WildEncounterAlterer;
 import battle.effect.generic.EffectNamesies;
-import battle.effect.generic.PokemonEffect;
-import battle.effect.generic.PokemonEffectNamesies;
-import battle.effect.generic.TerrainNamesies;
-import battle.effect.generic.WeatherEffect;
-import battle.effect.generic.WeatherNamesies;
+import battle.effect.generic.pokemon.PokemonEffect;
+import battle.effect.generic.pokemon.PokemonEffectNamesies;
+import battle.effect.generic.battle.terrain.TerrainNamesies;
+import battle.effect.generic.battle.weather.WeatherNamesies;
 import battle.effect.holder.AbilityHolder;
 import battle.effect.holder.ItemHolder;
 import battle.effect.status.Status;
