+++ src/test/ClassTest.java
@@ -8,7 +8,7 @@ import battle.effect.attack.ChangeTypeSource;
 import battle.effect.attack.MultiTurnMove;
 import battle.effect.attack.MultiTurnMove.ChargingMove;
 import battle.effect.attack.MultiTurnMove.RechargingMove;
-import battle.effect.generic.BattleEffect;
+import battle.effect.generic.battle.BattleEffect;
 import battle.effect.generic.EffectInterfaces.AbsorbDamageEffect;
 import battle.effect.generic.EffectInterfaces.AlwaysCritEffect;
 import battle.effect.generic.EffectInterfaces.ApplyDamageEffect;
@@ -78,9 +78,9 @@ import battle.effect.generic.EffectInterfaces.WeatherEliminatingEffect;
 import battle.effect.generic.EffectInterfaces.WeatherExtendingEffect;
 import battle.effect.generic.EffectInterfaces.WildEncounterAlterer;
 import battle.effect.generic.EffectInterfaces.WildEncounterSelector;
-import battle.effect.generic.PokemonEffect;
-import battle.effect.generic.PokemonEffectNamesies;
-import battle.effect.generic.TeamEffect;
+import battle.effect.generic.pokemon.PokemonEffect;
+import battle.effect.generic.pokemon.PokemonEffectNamesies;
+import battle.effect.generic.team.TeamEffect;
 import battle.effect.status.Status;
 import battle.effect.status.StatusCondition;
 import item.Item;
