+++ src/test/battle/AttackTest.java
@@ -5,15 +5,15 @@ import battle.attack.AttackNamesies;
 import battle.attack.Move;
 import battle.attack.MoveCategory;
 import battle.attack.MoveType;
-import battle.effect.generic.BattleEffect;
+import battle.effect.generic.battle.BattleEffect;
 import battle.effect.generic.Effect;
 import battle.effect.generic.EffectInterfaces.AlwaysCritEffect;
 import battle.effect.generic.EffectInterfaces.CritStageEffect;
 import battle.effect.generic.EffectInterfaces.SapHealthEffect;
 import battle.effect.generic.EffectInterfaces.SelfHealingMove;
 import battle.effect.generic.EffectNamesies;
-import battle.effect.generic.PokemonEffectNamesies;
-import battle.effect.generic.TeamEffectNamesies;
+import battle.effect.generic.pokemon.PokemonEffectNamesies;
+import battle.effect.generic.team.TeamEffectNamesies;
 import battle.effect.status.StatusCondition;
 import generator.update.MoveUpdater;
 import generator.update.MoveUpdater.MoveParser;
