+++ src/test/battle/EffectTest.java
@@ -7,8 +7,8 @@ import battle.attack.Move;
 import battle.effect.CastSource;
 import battle.effect.generic.Effect;
 import battle.effect.generic.EffectNamesies;
-import battle.effect.generic.PokemonEffectNamesies;
-import battle.effect.generic.TerrainNamesies;
+import battle.effect.generic.pokemon.PokemonEffectNamesies;
+import battle.effect.generic.battle.terrain.TerrainNamesies;
 import battle.effect.status.StatusCondition;
 import item.ItemNamesies;
 import org.junit.Assert;
