+++ src/test/battle/ModifierTest.java
@@ -4,11 +4,11 @@ import battle.attack.AttackNamesies;
 import battle.attack.Move;
 import battle.effect.CastSource;
 import battle.effect.attack.ChangeTypeSource;
-import battle.effect.generic.PokemonEffectNamesies;
-import battle.effect.generic.StandardBattleEffectNamesies;
-import battle.effect.generic.TeamEffectNamesies;
-import battle.effect.generic.TerrainNamesies;
-import battle.effect.generic.WeatherNamesies;
+import battle.effect.generic.pokemon.PokemonEffectNamesies;
+import battle.effect.generic.battle.StandardBattleEffectNamesies;
+import battle.effect.generic.team.TeamEffectNamesies;
+import battle.effect.generic.battle.terrain.TerrainNamesies;
+import battle.effect.generic.battle.weather.WeatherNamesies;
 import item.ItemNamesies;
 import org.junit.Assert;
 import org.junit.Test;
